Crear Informacion de la solicitud de viaje en fireabse database

+++ app/src/main/java/com/nain/cloneuber/models/FCMResponse.java
@@ -5,13 +5,13 @@ import java.util.ArrayList;
 public class FCMResponse {
     // mapeamos la respuesta cuando envaimos una notificacion
     // => https://www.site24x7.com/tools/json-to-java.html
-    private int multicast_id;
+    private long multicast_id;
     private int success;
     private int failure;
     private int canonical_ids;
     ArrayList<Object> results = new ArrayList<Object>();
 
-    public FCMResponse(int multicast_id, int success, int failure, int canonical_ids, ArrayList<Object> results) {
+    public FCMResponse(long multicast_id, int success, int failure, int canonical_ids, ArrayList<Object> results) {
         this.multicast_id = multicast_id;
         this.success = success;
         this.failure = failure;
@@ -19,11 +19,11 @@ public class FCMResponse {
         this.results = results;
     }
 
-    public int getMulticast_id() {
+    public long getMulticast_id() {
         return multicast_id;
     }
 
-    public void setMulticast_id(int multicast_id) {
+    public void setMulticast_id(long multicast_id) {
         this.multicast_id = multicast_id;
     }
